MAINT: Refactoring and other changes from code review
