+++ tests/test_benchmark.py
@@ -16,7 +16,7 @@ import numpy as np
 import pandas as pd
 from pandas.util.testing import assert_series_equal
 
-from zipline.data.historic_data_portal import HistoricDataPortal
+from zipline.data import HistoricDataPortal
 from zipline.errors import (
     BenchmarkAssetNotAvailableTooEarly,
     BenchmarkAssetNotAvailableTooLate,
