+++ zipline/data/__init__.py
@@ -3,7 +3,7 @@ from .loader import (
     load_prices_from_csv,
     load_prices_from_csv_folder,
 )
-from .data_portal import DataPortal
+from .data_portal import DataPortal, HistoricDataPortal
 
 
 __all__ = [
@@ -11,4 +11,5 @@ __all__ = [
     'load_prices_from_csv_folder',
     'loader',
     'DataPortal',
+    'HistoricDataPortal',
 ]
