+++ zipline/utils/run_algo.py
@@ -16,7 +16,7 @@ from trading_calendars import get_calendar
 
 from zipline.data import bundles, DataPortal
 from zipline.data.loader import load_market_data
-from zipline.data.historic_data_portal import HistoricDataPortal
+from zipline.data import HistoricDataPortal
 from zipline.finance import metrics
 from zipline.finance.trading import SimulationParameters
 from zipline.pipeline.data import USEquityPricing
@@ -73,7 +73,7 @@ def _run(handle_data,
          local_namespace,
          environ,
          blotter,
-         dataportal,
+         data_portal,
          benchmark_returns):
     """Run a backtest for the given algorithm.
 
@@ -171,23 +171,23 @@ def _run(handle_data,
         except ValueError as e:
             raise _RunAlgoError(str(e))
 
-    data = HistoricDataPortal(
-        bundle_data.asset_finder,
-        trading_calendar=trading_calendar,
-        first_trading_day=first_trading_day,
-        equity_minute_reader=bundle_data.equity_minute_bar_reader,
-        equity_daily_reader=bundle_data.equity_daily_bar_reader,
-        adjustment_reader=bundle_data.adjustment_reader,
-    )
-
-    if isinstance(dataportal, six.string_types):
-        if dataportal != "default":
+    if isinstance(data_portal, six.string_types):
+        if data_portal == 'default':
+            data = HistoricDataPortal(
+                bundle_data.asset_finder,
+                trading_calendar=trading_calendar,
+                first_trading_day=first_trading_day,
+                equity_minute_reader=bundle_data.equity_minute_bar_reader,
+                equity_daily_reader=bundle_data.equity_daily_bar_reader,
+                adjustment_reader=bundle_data.adjustment_reader,
+            )
+        else:
             try:
-                data = load(DataPortal, dataportal)
+                data = load(DataPortal, data_portal)
             except ValueError as e:
                 raise _RunAlgoError(str(e))
     else:
-        data = dataportal
+        data = data_portal
 
     perf = TradingAlgorithm(
         namespace=namespace,
@@ -388,6 +388,6 @@ def run_algorithm(start,
         local_namespace=False,
         environ=environ,
         blotter=blotter,
-        dataportal=dataportal,
+        data_portal=dataportal,
         benchmark_returns=benchmark_returns,
     )
